Resolve #617 Allow browsers other than Chrome
