+++ src/App.js
@@ -19,12 +19,8 @@ import { SnackbarProvider } from 'notistack'
 import NotifierComponent from './components/NotifierComponent'
 import IconButton from '@material-ui/core/IconButton'
 import CloseIcon from '@material-ui/icons/Close'
-import BrowserNotSupportedComponent from './components/BrowserNotSupportedComponent'
 import FAQContainer from './containers/FAQContainer'
-import { detect } from 'detect-browser'
 import { theme } from './styles/theme'
-
-const browser = detect()
 
 const userIsAuthenticated = connectedRouterRedirect({
   // The url to redirect user to if they fail
@@ -67,26 +63,13 @@ const componentStyle = {
   flexDirection: 'column'
 }
 
-function browserSupported () {
-  if (browser && browser.name === 'chrome') {
-    let v = browser.version.split('.')[0]
-    if (parseInt(v) >= 73) {
-      return true
-    }
-  }
-  return false
-}
-
 const DefaultLayout = ({ component: Component, ...rest }) => {
   return (
     <Route {...rest} render={matchProps => (
       <div style={defaultLayoutStyle}>
         <NaviBar {...matchProps} />
         <div style={componentStyle}>
-          {browserSupported()
-            ? <Component {...matchProps} />
-            : <BrowserNotSupportedComponent />
-          }
+          <Component {...matchProps} />
         </div>
         <NotifierComponent />
         <FAQContainer />
@@ -100,10 +83,7 @@ const LoginLayout = ({ component: Component, ...rest }) => {
   return (
     <Route {...rest} render={matchProps => (
       <div style={loginLayoutStyle}>
-        {browserSupported()
-          ? <Component {...matchProps} />
-          : <BrowserNotSupportedComponent />
-        }
+        <Component {...matchProps} />
         <NotifierComponent />
       </div>
     )} />
